+++ test/tiny_upgrader_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:tiny_upgrader/upgrader.dart';
 import 'package:tiny_upgrader/upgrader_platform_interface.dart';
 import 'package:tiny_upgrader/upgrader_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';